// api/src/modules/DAO/DAOdefs.d.ts
export type CompensatorDAO = {
    members: {
        supervisors: {
            list: ()=>Promise<Member[]>;
        }
    };
    slots:{
        getFromAvailability: (availability: Availability)=>Promise<Slot>;
    },
    availabilities: {
        update: (start:string, end:string, member_id:string, {state:AvailabilityState})=>Promise<void>;
    },
    interviewers: {
        list: ()=>Promise<Member[]>;
        listAvailableInDay: (timeInDay: string)=>Promise<Member[]>;
        leastNBusy: (N: number)=>Promise<Member[]>;
    },
    applications: {
        deleteSlot: (slot_id: number)=>Promise<void>;
        getBySlotId: (slot_id: number)=>Promise<Application>;
        getSlot: (id: number)=>Promise<Slot>;
    }
}